+++ src/pages/_app.tsx
@@ -1,48 +1,16 @@
-import React, { useEffect } from 'react';
+import React from 'react';
 
 import { ChakraProvider } from '@chakra-ui/react';
-import { useDVCClient, withDVCProvider } from '@devcycle/devcycle-react-sdk';
+import { withDVCProvider } from '@devcycle/devcycle-react-sdk';
+
+import { DVCDevToolProvider } from '../contexts/devcycledevtool';
 
 const App = ({ Component, pageProps }: any) => {
-  const dvcClient = useDVCClient();
-
-  useEffect(() => {
-    window.addEventListener(
-      'message',
-      function (e) {
-        // if (e.origin !== 'http://localhost:3001') {
-        //  // console.log("wrong");
-        //  return;
-        //}
-        try {
-          if (
-            e.data &&
-            e.data.type === '__DEVCYCLE_DEVTOOL_GLOBAL_HOOK_EVENT'
-          ) {
-            if (e.data.subType === 'GET_VARIABLES') {
-              console.log('react-app: GET_VARIABLES');
-              window.postMessage(
-                {
-                  type: '__DEVCYCLE_DEVTOOL_GLOBAL_HOOK_EVENT',
-                  subType: 'VARIABLES',
-                  data: JSON.stringify(dvcClient.allVariables())
-                },
-                '*'
-              );
-            }
-          }
-        } catch (err) {
-          console.log(e);
-          console.error(err);
-        }
-      },
-      false
-    );
-  }, [dvcClient]);
-
   return (
     <ChakraProvider>
-      <Component {...pageProps} />
+      <DVCDevToolProvider>
+        <Component {...pageProps} />
+      </DVCDevToolProvider>
     </ChakraProvider>
   );
 };
